requirements.txt
Flask>=3.0,<4
Flask-SQLAlchemy>=3.1,<4
SQLAlchemy>=2.0,<3
Werkzeug>=3.0,<4
Jinja2>=3.1,<4
python-dotenv>=1.0,<2
gunicorn>=22,<24
psycopg[binary]>=3.1,<4
scikit-learn>=1.5,<1.8
